requirements.txt
flask
flask-cors
google-generativeai
flask
flask-cors
google-generativeai
scikit-learn
python-dotenv
gunicorn
